+++ zellij-server/src/unit/screen_tests.rs
@@ -114,6 +114,7 @@ fn send_cli_action_to_server(
         route_action(
             action,
             client_id,
+            None,
             senders.clone(),
             capabilities,
             client_attributes.clone(),
@@ -1866,6 +1867,7 @@ pub fn send_cli_new_pane_action_with_default_parameters() {
         plugin: None,
         cwd: None,
         floating: false,
+        in_place: false,
         name: None,
         close_on_exit: false,
         start_suspended: false,
@@ -1903,6 +1905,7 @@ pub fn send_cli_new_pane_action_with_split_direction() {
         plugin: None,
         cwd: None,
         floating: false,
+        in_place: false,
         name: None,
         close_on_exit: false,
         start_suspended: false,
@@ -1940,6 +1943,7 @@ pub fn send_cli_new_pane_action_with_command_and_cwd() {
         plugin: None,
         cwd: Some("/some/folder".into()),
         floating: false,
+        in_place: false,
         name: None,
         close_on_exit: false,
         start_suspended: false,
@@ -1976,6 +1980,7 @@ pub fn send_cli_edit_action_with_default_parameters() {
         direction: None,
         line_number: None,
         floating: false,
+        in_place: false,
         cwd: None,
     };
     send_cli_action_to_server(&session_metadata, cli_edit_action, client_id);
@@ -2009,6 +2014,7 @@ pub fn send_cli_edit_action_with_line_number() {
         direction: None,
         line_number: Some(100),
         floating: false,
+        in_place: false,
         cwd: None,
     };
     send_cli_action_to_server(&session_metadata, cli_edit_action, client_id);
@@ -2042,6 +2048,7 @@ pub fn send_cli_edit_action_with_split_direction() {
         direction: Some(Direction::Down),
         line_number: None,
         floating: false,
+        in_place: false,
         cwd: None,
     };
     send_cli_action_to_server(&session_metadata, cli_edit_action, client_id);
@@ -2567,6 +2574,7 @@ pub fn send_cli_launch_or_focus_plugin_action() {
     );
     let cli_action = CliAction::LaunchOrFocusPlugin {
         floating: true,
+        in_place: false,
         move_to_focused_tab: true,
         url: url::Url::parse("file:/path/to/fake/plugin").unwrap(),
         configuration: Default::default(),
@@ -2625,6 +2633,7 @@ pub fn send_cli_launch_or_focus_plugin_action_when_plugin_is_already_loaded() {
     );
     let cli_action = CliAction::LaunchOrFocusPlugin {
         floating: true,
+        in_place: false,
         move_to_focused_tab: true,
         url: url::Url::parse("file:/path/to/fake/plugin").unwrap(),
         configuration: Default::default(),