+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+* feat(panes): start panes/editors/commands/plugins in-place (https://github.com/zellij-org/zellij/pull/2795)
+
+## [0.38.2] - 2023-09-15
+* fix(terminal): wrap lines in alternate screen mode when adding characters (https://github.com/zellij-org/zellij/pull/2789)
+* fix(utils): validate session name (https://github.com/zellij-org/zellij/pull/2607)
 
 ## [0.38.1] - 2023-08-31
 * refactor(server): remove unnecessary mut (https://github.com/zellij-org/zellij/pull/2735)